Move `CopyUserGroup` command
