+++ wcfsetup/install/files/lib/acp/action/UserGroupCopyAction.class.php
@@ -15,7 +15,7 @@ use wcf\system\form\builder\container\FormContainer;
 use wcf\system\form\builder\field\BooleanFormField;
 use wcf\system\form\builder\Psr15DialogForm;
 use wcf\system\request\LinkHandler;
-use wcf\system\user\group\command\CopyUserGroup;
+use wcf\command\user\group\CopyUserGroup;
 use wcf\system\WCF;
 
 /**
